ok, this is a mid release commit. i cleaned the code a lot. as a result the entire module should be more stable now and more easy to maintain. this version does not implement the memory management code, i still have in my mind.
because everytime a node is retured a new scalar is created, this version
leaks as hell (probably all earlier versions did as well). this will
change with one of the next commits.

one other thing left to do, is to make a switch for toString available, so
a user may decide wether to recieve the string as UTF8 or as the DOM
encoding.

Modified Files:
LibXML.xs
   + over all cleanup
   + recognize strings as utf8 or document encoding (only perl 5.6.x+)
   + all string returned are utf8 (perl flag is set correctly)
   ++ removed a lot redundand code (almost all SvREFCNT*() calls)
   - $ORIGINAL_STRING code (was a bad idea, sorry for adding it)

MANIFEST
   + perl-libxml-mm.h
   + perl-libxml-mm.c

Makefile.PL
   + checks the utf8 capability of perl (-DHAVE_UTF8 is set if perl is at least
     5.6.0)

        LibXML.pm dom.c dom.h
 	t/01basic.t t/02parsestring.t t/05dombasic.t t/06nodetypes.t
 	t/07nodelist.t t/10ns.t
        (---)

Added Files:
perl-libxml-mm.c perl-libxml-mm.h
    + implements the common logic of the MM (Perl <-> C layer)