Fix handling of DTD nodes
When removing a DTD node, it must not be added to a document fragment.
Otherwise it will leak. Also make sure that removed attributes aren't
added to document fragments.

When inserting a DTD node, make sure that the 'intSubset' member of the
document struct is updated and the previous DTD is unlinked and freed.
This fixes memory leaks and potential segfaults. See RT #80521:

https://rt.cpan.org/Ticket/Display.html?id=80521

Also fix a memory leak in $node->removeChildNodes and make sure that
the correct owner is set when inserting nodes from other documents via
$node->insertBefore ro $node->insertAfter.