Fix double free when calling $node->addSibling with text nodes
In libxml2?s tree.c xmlAddSibling has an optimization where if the
existing node is text and you appendSibling a text node it will instead
append the text onto the existing sibling then xmlFree the new sibling.

Handle this special case by copying the text node before adding it.

Reported in RT #94149
https://rt.cpan.org/Ticket/Display.html?id=94149

Thanks to Jeff Trout for the report.